Split up lib.zookeeper
