+++ infrastructure/dns_server/main.py
@@ -45,7 +45,8 @@ from lib.defines import (
     SCION_DNS_PORT,
     SIBRA_SERVICE,
 )
-from lib.zookeeper import ZkNoConnection, Zookeeper
+from lib.zk.errors import ZkNoConnection
+from lib.zk.zk import Zookeeper
 
 
 class SCIONDnsServer(SCIONElement):
