+++ lib/zk/errors.py
@@ -1,0 +1,39 @@
+# Copyright 2016 ETH Zurich
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+:mod:`errors` --- lib.zk Errors
+===============================
+"""
+# SCION
+from lib.errors import SCIONBaseError
+
+
+class ZkBaseError(SCIONBaseError):
+    """Base exception class for all lib.zookeeper exceptions."""
+    pass
+
+
+class ZkNoConnection(ZkBaseError):
+    """No connection to Zookeeper."""
+    pass
+
+
+class ZkNoNodeError(ZkBaseError):
+    """A node doesn't exist."""
+    pass
+
+
+class ZkRetryLimit(ZkBaseError):
+    """Operation hit retry limit."""
+    pass
