+++ sphinx-doc/lib/zk/cache.rst
@@ -1,4 +1,4 @@
-.. automodule:: lib.zookeeper
+.. automodule:: lib.zk.cache
    :special-members: __init__
    :members:
    :member-order: bysource
