+++ sphinx-doc/lib/zk/errors.rst
@@ -1,0 +1,4 @@
+.. automodule:: lib.zk.errors
+   :special-members: __init__
+   :members:
+   :member-order: bysource
