+++ sphinx-doc/lib/zk/index.rst
@@ -1,0 +1,8 @@
+SCION Zookeeper Library
+=======================
+
+.. toctree::
+  :maxdepth: 3
+  :glob:
+
+  *
