+++ sphinx-doc/lib/zk/party.rst
@@ -1,0 +1,4 @@
+.. automodule:: lib.zk.party
+   :special-members: __init__
+   :members:
+   :member-order: bysource
