+++ sphinx-doc/lib/zk/zk.rst
@@ -1,0 +1,4 @@
+.. automodule:: lib.zk.zk
+   :special-members: __init__
+   :members:
+   :member-order: bysource
