+++ test/lib/zk/zk_test.py
@@ -1,4 +1,4 @@
-# Copyright 2015 ETH Zurich
+# Copyright 2016 ETH Zurich
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-:mod:`lib_zookeeper_test` --- lib.zookeeper unit tests
-======================================================
+:mod:`zk_test` --- lib.zk.zk unit tests
+=======================================
 """
 # Stdlib
 import logging
@@ -27,27 +27,20 @@ from kazoo.exceptions import (
     ConnectionLoss,
     LockTimeout,
     NoNodeError,
-    NodeExistsError,
     SessionExpiredError,
 )
 from kazoo.handlers.threading import KazooTimeoutError
 
 # SCION
 from lib.thread import thread_safety_net
-from lib.zookeeper import (
-    ZkNoConnection,
-    ZkNoNodeError,
-    ZkParty,
-    ZkRetryLimit,
-    ZkSharedCache,
-    Zookeeper,
-)
-from test.testcommon import SCIONTestError, assert_these_calls, create_mock
+from lib.zk.errors import ZkNoConnection
+from lib.zk.zk import ZkRetryLimit, Zookeeper
+from test.testcommon import SCIONTestError, create_mock
 
 
 class BaseZookeeper(object):
     """
-    Base class for lib.zookeeper.Zookeeper unit tests
+    Base class for lib.zk.zk.Zookeeper unit tests
     """
     default_args = ["1-2", "srvtype", "srvid"]
     default_hosts = ["host1:9521", "host2:339"]
@@ -59,13 +52,13 @@ class BaseZookeeper(object):
 
 class TestZookeeperInit(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.__init__
+    Unit tests for lib.zk.zk.Zookeeper.__init__
     """
-    @patch("lib.zookeeper.Zookeeper._kazoo_start", autospec=True)
-    @patch("lib.zookeeper.Zookeeper._setup_state_listener", autospec=True)
-    @patch("lib.zookeeper.Zookeeper._kazoo_setup", autospec=True)
-    @patch("lib.zookeeper.queue.Queue", autospec=True)
-    @patch("lib.zookeeper.threading.Event", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._kazoo_start", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._setup_state_listener", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._kazoo_setup", autospec=True)
+    @patch("lib.zk.zk.queue.Queue", autospec=True)
+    @patch("lib.zk.zk.threading.Event", autospec=True)
     def test_full(self, event, queue, ksetup, listener, kstart):
         # Setup and call
         event.side_effect = ["event0", "event1"]
@@ -89,11 +82,11 @@ class TestZookeeperInit(BaseZookeeper):
         listener.assert_called_once_with(inst)
         kstart.assert_called_once_with(inst)
 
-    @patch("lib.zookeeper.Zookeeper._kazoo_start", autospec=True)
-    @patch("lib.zookeeper.Zookeeper._setup_state_listener", autospec=True)
-    @patch("lib.zookeeper.Zookeeper._kazoo_setup", autospec=True)
-    @patch("lib.zookeeper.threading.Semaphore", autospec=True)
-    @patch("lib.zookeeper.threading.Event", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._kazoo_start", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._setup_state_listener", autospec=True)
+    @patch("lib.zk.zk.Zookeeper._kazoo_setup", autospec=True)
+    @patch("lib.zk.zk.threading.Semaphore", autospec=True)
+    @patch("lib.zk.zk.threading.Event", autospec=True)
     def test_defaults(self, event, semaphore, ksetup, listener, kstart):
         # Setup and call
         inst = self._init_basic_setup()
@@ -105,12 +98,12 @@ class TestZookeeperInit(BaseZookeeper):
 
 class TestZookeeperKazooSetup(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._kazoo_setup
+    Unit tests for lib.zk.zk.Zookeeper._kazoo_setup
     """
-    @patch("lib.zookeeper.KazooClient", autospec=True)
-    @patch("lib.zookeeper.logging.getLogger", autospec=True)
-    @patch("lib.zookeeper.KazooRetry", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.KazooClient", autospec=True)
+    @patch("lib.zk.zk.logging.getLogger", autospec=True)
+    @patch("lib.zk.zk.KazooRetry", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test(self, init, kretry, getlogger, kclient):
         # Setup
         inst = self._init_basic_setup()
@@ -131,10 +124,10 @@ class TestZookeeperKazooSetup(BaseZookeeper):
 
 class TestZookeeperKazooStart(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._kazoo_start
+    Unit tests for lib.zk.zk.Zookeeper._kazoo_start
     """
-    @patch("lib.zookeeper.logging", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.logging", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test(self, init, logging_):
         # Setup
         inst = self._init_basic_setup()
@@ -144,9 +137,9 @@ class TestZookeeperKazooStart(BaseZookeeper):
         # Tests
         inst.kazoo.start.assert_called_once_with()
 
-    @patch("lib.zookeeper.kill_self", autospec=True)
-    @patch("lib.zookeeper.logging", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.kill_self", autospec=True)
+    @patch("lib.zk.zk.logging", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_timeout(self, init, logging_, kill_self):
         # Setup
         inst = self._init_basic_setup()
@@ -160,10 +153,10 @@ class TestZookeeperKazooStart(BaseZookeeper):
 
 class TestZookeeperSetupStateListener(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._setup_state_listener
+    Unit tests for lib.zk.zk.Zookeeper._setup_state_listener
     """
-    @patch("lib.zookeeper.threading.Thread", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.threading.Thread", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test(self, init, thread):
         # Setup
         inst = self._init_basic_setup()
@@ -181,9 +174,9 @@ class TestZookeeperSetupStateListener(BaseZookeeper):
 
 class TestZookeeperStateListener(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._state_listener
+    Unit tests for lib.zk.zk.Zookeeper._state_listener
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test(self, init):
         # Setup
         inst = self._init_basic_setup()
@@ -198,9 +191,9 @@ class TestZookeeperStateListener(BaseZookeeper):
 
 class TestZookeeperStateHandler(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._state_handler
+    Unit tests for lib.zk.zk.Zookeeper._state_handler
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_flapping(self, init):
         # Setup
         inst = self._init_basic_setup()
@@ -215,7 +208,7 @@ class TestZookeeperStateHandler(BaseZookeeper):
         inst._state_events.get.assert_has_calls([call()] * 2)
         ntools.assert_false(inst._state_connected.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check(self, old_state, new_state, init):
         # Setup
         inst = self._init_basic_setup()
@@ -261,7 +254,7 @@ class TestZookeeperStateHandler(BaseZookeeper):
 
 class TestZookeeperStateConnected(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._state_connected
+    Unit tests for lib.zk.zk.Zookeeper._state_connected
     """
     def _setup(self):
         inst = self._init_basic_setup()
@@ -277,7 +270,7 @@ class TestZookeeperStateConnected(BaseZookeeper):
         inst._on_connect = None
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_basic(self, init):
         inst = self._setup()
         # Call
@@ -288,7 +281,7 @@ class TestZookeeperStateConnected(BaseZookeeper):
         inst._parties["/pathb"].autojoin.assert_called_once_with()
         inst._connected.set.assert_called_once_with()
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_connloss(self, init):
         inst = self._setup()
         inst.ensure_path.side_effect = ZkNoConnection
@@ -297,7 +290,7 @@ class TestZookeeperStateConnected(BaseZookeeper):
         # Tests
         ntools.assert_false(inst._connected.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_on_connect(self, init):
         inst = self._setup()
         inst._on_connect = create_mock()
@@ -309,13 +302,13 @@ class TestZookeeperStateConnected(BaseZookeeper):
 
 class TestZookeeperStateDisconnected(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper._state_suspended AND
-    lib.zookeeper.Zookeeper._state_lost
+    Unit tests for lib.zk.zk.Zookeeper._state_suspended AND
+    lib.zk.zk.Zookeeper._state_lost
 
     _state_suspended and _state_lost currently have almost identical code, so
     test them both in the same way.
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check(self, f_name, init, test_callback=False):
         inst = self._init_basic_setup()
         inst._connected = create_mock(["clear"])
@@ -340,9 +333,9 @@ class TestZookeeperStateDisconnected(BaseZookeeper):
 
 class TestZookeeperIsConnected(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.is_connected
+    Unit tests for lib.zk.zk.Zookeeper.is_connected
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check(self, connected, init):
         inst = self._init_basic_setup()
         inst._connected = create_mock(["is_set"])
@@ -359,9 +352,9 @@ class TestZookeeperIsConnected(BaseZookeeper):
 
 class TestZookeeperWaitConnected(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.wait_connected
+    Unit tests for lib.zk.zk.Zookeeper.wait_connected
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_connected(self, init):
         inst = self._init_basic_setup()
         inst.is_connected = create_mock()
@@ -370,13 +363,14 @@ class TestZookeeperWaitConnected(BaseZookeeper):
         # Tests
         inst.is_connected.assert_called_once_with()
 
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
-    def test_no_timeout(self, init, get_time):
+    @patch("lib.zk.zk.time.time", autospec=True)
+    @patch("lib.zk.zk.logging.debug", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
+    def test_no_timeout(self, init, _, time_):
         inst = self._init_basic_setup()
         inst.is_connected = create_mock()
         inst.is_connected.return_value = False
-        get_time.side_effect = [0, 10, 20]
+        time_.side_effect = [0, 10, 20]
         inst._connected = create_mock(["wait"])
         inst._connected.wait.side_effect = [False, True]
         # Call
@@ -385,13 +379,14 @@ class TestZookeeperWaitConnected(BaseZookeeper):
         inst._connected.wait.assert_has_calls([call(timeout=10.0)] * 2)
         ntools.eq_(inst._connected.wait.call_count, 2)
 
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
-    def test_timeout_success(self, init, get_time):
+    @patch("lib.zk.zk.time.time", autospec=True)
+    @patch("lib.zk.zk.logging.debug", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
+    def test_timeout_success(self, init, _, time_):
         inst = self._init_basic_setup()
         inst.is_connected = create_mock()
         inst.is_connected.return_value = False
-        get_time.side_effect = [0, 10, 20]
+        time_.side_effect = [0, 10, 20]
         inst._connected = create_mock(["wait"])
         inst._connected.wait.side_effect = [False, True]
         # Call
@@ -401,13 +396,14 @@ class TestZookeeperWaitConnected(BaseZookeeper):
                                                call(timeout=5.0)])
         ntools.eq_(inst._connected.wait.call_count, 2)
 
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
-    def test_timeout_fail(self, init, get_time):
+    @patch("lib.zk.zk.time.time", autospec=True)
+    @patch("lib.zk.zk.logging.debug", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
+    def test_timeout_fail(self, init, _, time_):
         inst = self._init_basic_setup()
         inst.is_connected = create_mock()
         inst.is_connected.return_value = False
-        get_time.side_effect = [0, 10, 20]
+        time_.side_effect = [0, 10, 20]
         inst._connected = create_mock(["wait"])
         inst._connected.wait.side_effect = [False, False]
         # Call
@@ -418,7 +414,7 @@ class TestZookeeperWaitConnected(BaseZookeeper):
 
 class TestZookeeperEnsurePath(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.ensure_path
+    Unit tests for lib.zk.zk.Zookeeper.ensure_path
     """
     def _setup(self):
         inst = self._init_basic_setup()
@@ -426,7 +422,7 @@ class TestZookeeperEnsurePath(BaseZookeeper):
         inst.kazoo = create_mock(["ensure_path"])
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_basic(self, init):
         # Setup
         inst = self._setup()
@@ -435,7 +431,7 @@ class TestZookeeperEnsurePath(BaseZookeeper):
         # Tests
         inst.kazoo.ensure_path.assert_called_once_with("/prefix/pathness")
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_abs(self, init):
         # Setup
         inst = self._setup()
@@ -444,7 +440,7 @@ class TestZookeeperEnsurePath(BaseZookeeper):
         # Tests
         inst.kazoo.ensure_path.assert_called_once_with("/path/to/stuff")
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check_error(self, excp, init):
         # Setup
         inst = self._setup()
@@ -459,7 +455,7 @@ class TestZookeeperEnsurePath(BaseZookeeper):
 
 class TestZookeeperPartySetup(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.party_setup
+    Unit tests for lib.zk.zk.Zookeeper.party_setup
     """
     def _setup(self, connected=True):
         inst = self._init_basic_setup()
@@ -472,7 +468,7 @@ class TestZookeeperPartySetup(BaseZookeeper):
         inst._parties = {}
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_not_connected(self, init):
         inst = self._setup(connected=False)
         # Call
@@ -480,8 +476,8 @@ class TestZookeeperPartySetup(BaseZookeeper):
         # Tests
         inst.is_connected.assert_called_once_with()
 
-    @patch("lib.zookeeper.ZkParty", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.ZkParty", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_basic(self, init, zkparty):
         inst = self._setup()
         # Call
@@ -492,8 +488,8 @@ class TestZookeeperPartySetup(BaseZookeeper):
                                         inst._srv_id, True)
         ntools.eq_(inst._parties, {"/prefix/party": zkparty.return_value})
 
-    @patch("lib.zookeeper.ZkParty", autospec=True)
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.ZkParty", autospec=True)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_full(self, init, zkparty):
         inst = self._setup()
         # Call
@@ -505,7 +501,7 @@ class TestZookeeperPartySetup(BaseZookeeper):
 
 class TestZookeeperGetLock(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.get_lock
+    Unit tests for lib.zk.zk.Zookeeper.get_lock
     """
     def _setup(self, have_lock):
         inst = self._init_basic_setup()
@@ -519,7 +515,7 @@ class TestZookeeperGetLock(BaseZookeeper):
         inst._lock = create_mock(["set"])
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_full(self, init):
         inst = self._setup((True,))
         inst._zk_lock = None
@@ -538,7 +534,7 @@ class TestZookeeperGetLock(BaseZookeeper):
         inst._zk_lock.acquire.assert_called_once_with(timeout="lock t/o")
         inst._lock.set.assert_called_once_with()
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_have_lock(self, init):
         inst = self._setup((True,))
         # Call
@@ -546,7 +542,7 @@ class TestZookeeperGetLock(BaseZookeeper):
         # Tests
         inst.have_lock.assert_called_once_with()
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_acquire_fail(self, init):
         inst = self._setup((False, False))
         inst._zk_lock.acquire.return_value = False
@@ -557,7 +553,7 @@ class TestZookeeperGetLock(BaseZookeeper):
         inst._zk_lock.acquire.assert_called_once_with(timeout=None)
         ntools.assert_false(inst._lock.set.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_acquire_timeout(self, init):
         inst = self._setup((False, False))
         inst._zk_lock.acquire.side_effect = LockTimeout
@@ -566,7 +562,7 @@ class TestZookeeperGetLock(BaseZookeeper):
         # Tests
         ntools.assert_false(inst._lock.set.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check_exception(self, excp, init):
         inst = self._setup((False,))
         inst._zk_lock.acquire.side_effect = excp
@@ -580,9 +576,9 @@ class TestZookeeperGetLock(BaseZookeeper):
 
 class TestZookeeperReleaseLock(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.release_lock
+    Unit tests for lib.zk.zk.Zookeeper.release_lock
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_no_lock(self, init):
         inst = self._init_basic_setup()
         inst._lock = create_mock(["clear"])
@@ -594,7 +590,7 @@ class TestZookeeperReleaseLock(BaseZookeeper):
         inst._lock.clear.assert_called_once_with()
         ntools.assert_false(inst.is_connected.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_not_connected(self, init):
         inst = self._init_basic_setup()
         inst._lock = create_mock(["clear"])
@@ -609,7 +605,7 @@ class TestZookeeperReleaseLock(BaseZookeeper):
         inst.is_connected.assert_called_once_with()
         ntools.assert_false(inst._zk_lock.is_acquired)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_release(self, init):
         inst = self._init_basic_setup()
         inst._lock = create_mock(["clear"])
@@ -620,7 +616,7 @@ class TestZookeeperReleaseLock(BaseZookeeper):
         # Tests
         inst._zk_lock.release.assert_called_once_with()
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check_exception(self, exception, init):
         inst = self._init_basic_setup()
         inst._lock = create_mock(["clear"])
@@ -641,7 +637,7 @@ class TestZookeeperReleaseLock(BaseZookeeper):
 
 class TestZookeeperHaveLock(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.have_lock
+    Unit tests for lib.zk.zk.Zookeeper.have_lock
     """
     def _setup(self, connected, l_epoch, c_epoch, lock_is_set):
         inst = self._init_basic_setup()
@@ -654,7 +650,7 @@ class TestZookeeperHaveLock(BaseZookeeper):
         inst.release_lock = create_mock()
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_have(self, init):
         inst = self._setup(True, 1, 1, True)
         # Call
@@ -664,7 +660,7 @@ class TestZookeeperHaveLock(BaseZookeeper):
         inst._lock.is_set.assert_called_once_with()
         ntools.assert_false(inst.release_lock.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check_not_have(self, connected, l_epoch, c_epoch, lock_is_set, init):
         inst = self._setup(connected, l_epoch, c_epoch, lock_is_set)
         # Call
@@ -683,9 +679,9 @@ class TestZookeeperHaveLock(BaseZookeeper):
 
 class TestZookeeperWaitLock(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.wait_lock
+    Unit tests for lib.zk.zk.Zookeeper.wait_lock
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test(self, init):
         inst = self._init_basic_setup()
         inst._lock = create_mock(["wait"])
@@ -697,7 +693,7 @@ class TestZookeeperWaitLock(BaseZookeeper):
 
 class TestZookeeperGetLockHolder(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.get_lock_holder
+    Unit tests for lib.zk.zk.Zookeeper.get_lock_holder
     """
     def _setup(self, contenders):
         inst = self._init_basic_setup()
@@ -705,7 +701,7 @@ class TestZookeeperGetLockHolder(BaseZookeeper):
         inst._zk_lock.contenders.return_value = contenders
         return inst
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_without_contenders(self, init):
         inst = self._setup([])
         # Call
@@ -713,13 +709,13 @@ class TestZookeeperGetLockHolder(BaseZookeeper):
         # Tests
         inst._zk_lock.contenders.assert_called_once_with()
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_with_contenders(self, init):
         inst = self._setup(["12\x0034\x0056", "78\x0090\x00ab"])
         # Call
         ntools.eq_(inst.get_lock_holder(), ["12", "34", "56"])
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def _check_error(self, excp, init):
         # Setup
         inst = self._setup([])
@@ -734,9 +730,9 @@ class TestZookeeperGetLockHolder(BaseZookeeper):
 
 class TestZookeeperRetry(BaseZookeeper):
     """
-    Unit tests for lib.zookeeper.Zookeeper.retry
+    Unit tests for lib.zk.zk.Zookeeper.retry
     """
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_basic(self, init):
         inst = self._init_basic_setup()
         inst.wait_connected = create_mock()
@@ -748,7 +744,7 @@ class TestZookeeperRetry(BaseZookeeper):
         inst.wait_connected.assert_called_once_with(timeout=5.4)
         f.assert_called_once_with("arg1", kwarg1="k")
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_no_conn(self, init):
         inst = self._init_basic_setup()
         inst.wait_connected = create_mock()
@@ -761,7 +757,7 @@ class TestZookeeperRetry(BaseZookeeper):
         ntools.eq_(inst.wait_connected.call_count, 5)
         ntools.assert_false(f.called)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_no_retries(self, init):
         inst = self._init_basic_setup()
         inst.wait_connected = create_mock()
@@ -772,7 +768,7 @@ class TestZookeeperRetry(BaseZookeeper):
         # Tests
         inst.wait_connected.assert_called_once_with(timeout=10.0)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_inf_retries(self, init):
         inst = self._init_basic_setup()
         inst.wait_connected = create_mock()
@@ -785,7 +781,7 @@ class TestZookeeperRetry(BaseZookeeper):
         inst.wait_connected.assert_has_calls([call(timeout=10.0)] * 21)
         ntools.eq_(inst.wait_connected.call_count, 21)
 
-    @patch("lib.zookeeper.Zookeeper.__init__", autospec=True, return_value=None)
+    @patch("lib.zk.zk.Zookeeper.__init__", autospec=True, return_value=None)
     def test_conn_drop(self, init):
         inst = self._init_basic_setup()
         inst.wait_connected = create_mock()
@@ -797,410 +793,6 @@ class TestZookeeperRetry(BaseZookeeper):
         # Tests
         inst.wait_connected.assert_has_calls([call(timeout=10.0)] * 2)
         ntools.eq_(inst.wait_connected.call_count, 2)
-
-
-class TestZkPartyInit(object):
-    """
-    Unit tests for lib.zookeeper.ZkParty.__init__
-    """
-    @patch("lib.zookeeper.ZkParty.autojoin", autospec=True)
-    def test_basic(self, autojoin):
-        zk = create_mock(["Party"])
-        # Call
-        p = ZkParty(zk, "path", "id", "autojoin")
-        # Tests
-        ntools.eq_(p._autojoin, "autojoin")
-        ntools.eq_(p._path, "path")
-        zk.Party.assert_called_once_with("path", "id")
-        ntools.eq_(p._party, zk.Party.return_value)
-        autojoin.assert_called_once_with(p)
-
-    @patch("lib.zookeeper.ZkParty.autojoin", autospec=True)
-    def _check_error(self, excp, autojoin):
-        zk = create_mock(["Party"])
-        zk.Party.side_effect = excp
-        # Call
-        ntools.assert_raises(ZkNoConnection, ZkParty, zk, "path",
-                             "id", True)
-
-    def test_error(self):
-        for excp in ConnectionLoss, SessionExpiredError:
-            yield self._check_error, excp
-
-
-class TestZkPartyJoin(object):
-    """
-    Unit tests for lib.zookeeper.ZkParty.join
-    """
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def test_basic(self, init):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._party = create_mock(["join"])
-        p.list = create_mock()
-        # Call
-        p.join()
-        # Tests
-        p._party.join.assert_called_once_with()
-
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def _check_error(self, excp, init):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._party = create_mock(["join"])
-        p._party.join.side_effect = excp
-        # Call
-        ntools.assert_raises(ZkNoConnection, p.join)
-
-    def test_error(self):
-        for excp in ConnectionLoss, SessionExpiredError:
-            yield self._check_error, excp
-
-
-class TestZkPartyAutoJoin(object):
-    """
-    Unit tests for lib.zookeeper.ZkParty.autojoin
-    """
-    @patch("lib.zookeeper.ZkParty.list", autospec=True)
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def test_auto(self, init, list_):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._autojoin = True
-        p.join = create_mock()
-        p._path = "path"
-        # Call
-        p.autojoin()
-        # Tests
-        p.join.assert_called_once_with()
-
-    @patch("lib.zookeeper.ZkParty.list", autospec=True)
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def test_noauto(self, init, list_):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._autojoin = False
-        p.join = create_mock()
-        p._path = "path"
-        # Call
-        p.autojoin()
-        # Tests
-        ntools.assert_false(p.join.called)
-
-
-class TestZkPartyList(object):
-    """
-    Unit tests for lib.zookeeper.ZkParty.list
-    """
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def test_basic(self, init):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._party = MagicMock(spec_set=["__iter__"])
-        p._party.__iter__.return_value = [1, 2, 3]
-        # Call
-        ntools.eq_(p.list(), {1, 2, 3})
-
-    @patch("lib.zookeeper.ZkParty.__init__", autospec=True, return_value=None)
-    def _check_error(self, excp, init):
-        p = ZkParty("zk", "path", "id", "autojoin")
-        p._party = create_mock(["__iter__"])
-        p._party.__iter__.side_effect = excp
-        # Call
-        ntools.assert_raises(ZkNoConnection, p.list)
-
-    def test_error(self):
-        for excp in ConnectionLoss, SessionExpiredError:
-            yield self._check_error, excp
-
-
-class TestZkSharedCacheStore(object):
-    """
-    Unit tests for lib.zookeeper.ZkSharedCache.store
-    """
-    def _setup(self):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._zk = create_mock(["is_connected"])
-        inst._kazoo = create_mock(["create", "set"])
-        inst._incoming_entries = create_mock(["append"])
-        return inst
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_not_connected(self, init):
-        inst = self._setup()
-        inst._zk.is_connected.return_value = False
-        # Call
-        ntools.assert_raises(ZkNoConnection, inst.store, 'n', 'v')
-        # Tests
-        inst._zk.is_connected.assert_called_once_with()
-
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_set(self, init, get_time):
-        inst = self._setup()
-        # Call
-        inst.store('n', 'v')
-        # Tests
-        inst._kazoo.set.assert_called_once_with("/path/n", "v")
-        ntools.assert_false(inst._kazoo.create.called)
-        inst._incoming_entries.append.assert_called_once_with(
-            ("n", get_time.return_value))
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def _check_set_conn_loss(self, excp, init):
-        inst = self._setup()
-        inst._kazoo.set.side_effect = excp
-        # Call
-        ntools.assert_raises(ZkNoConnection, inst.store, 'n', 'v')
-
-    def test_set_conn_loss(self):
-        for excp in ConnectionLoss, SessionExpiredError:
-            yield self._check_set_conn_loss, excp
-
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_create(self, init, get_time):
-        inst = self._setup()
-        inst._kazoo.set.side_effect = NoNodeError
-        # Call
-        inst.store('n', 'v')
-        # Tests
-        inst._kazoo.create.assert_called_once_with("/path/n", "v",
-                                                   makepath=True)
-        inst._incoming_entries.append.assert_called_once_with(
-            ("n", get_time.return_value))
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_suddenly_exists(self, init):
-        inst = self._setup()
-        inst._kazoo.set.side_effect = NoNodeError
-        inst._kazoo.create.side_effect = NodeExistsError
-        # Call
-        inst.store('n', 'v')
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def _check_create_conn_loss(self, excp, init):
-        inst = self._setup()
-        inst._kazoo.set.side_effect = NoNodeError
-        inst._kazoo.create.side_effect = excp
-        # Call
-        ntools.assert_raises(ZkNoConnection, inst.store, 'n', 'v')
-
-    def test_create_conn_loss(self):
-        for excp in ConnectionLoss, SessionExpiredError:
-            yield self._check_create_conn_loss, excp
-
-
-class TestZkSharedCacheProcess(object):
-    """
-    Unit tests for lib.zookeeper.ZkSharedCache.process
-    """
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_not_connected(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._zk = create_mock(["is_connected"])
-        inst._zk.is_connected.return_value = False
-        # Call
-        ntools.assert_raises(ZkNoConnection, inst.process)
-        # Tests
-        inst._zk.is_connected.assert_called_once_with()
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_full(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._zk = create_mock(["conn_epoch", "is_connected"])
-        inst._incoming_entries = create_mock(["__bool__", "popleft"])
-        inst._incoming_entries.__bool__.side_effect = True, True, False
-        inst._incoming_entries.popleft.side_effect = ("inc0", 1), ("inc1", 0)
-        inst._entries = {"inc0": 0, "old0": 0}
-        inst._list_entries = create_mock()
-        inst._list_entries.return_value = "inc0", "inc1", "new0"
-        inst._handle_entries = create_mock()
-        inst._path = "/path"
-        # Call
-        inst.process()
-        # Tests
-        ntools.eq_(inst._entries, {"inc0": 0, "inc1": 0})
-        inst._handle_entries.assert_called_once_with({"new0"})
-
-
-class TestZkSharedCacheGet(object):
-    """
-    Unit tests for lib.zookeeper.ZkSharedCache._get
-    """
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_success(self, init, get_time):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get"])
-        inst._kazoo.get.return_value = ("data", "meta")
-        inst._entries = create_mock(["setdefault"])
-        # Call
-        ntools.eq_(inst._get("name"), "data")
-        # Tests
-        inst._kazoo.get.assert_called_once_with("/path/name")
-        inst._entries.setdefault.assert_called_once_with(
-            "name", get_time.return_value)
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_no_entry(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get"])
-        inst._kazoo.get.side_effect = NoNodeError
-        inst._entries = create_mock(["pop"])
-        # Call
-        ntools.assert_raises(ZkNoNodeError, inst._get, "name")
-        # Tests
-        inst._kazoo.get.assert_called_once_with("/path/name")
-        inst._entries.pop.assert_called_once_with("name", None)
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def _check_exception(self, excp, expected, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get"])
-        inst._kazoo.get.side_effect = excp
-        # Call
-        ntools.assert_raises(expected, inst._get, "name")
-
-    def test_exceptions(self):
-        for excp, expected in (
-            (ConnectionLoss, ZkNoConnection),
-            (SessionExpiredError, ZkNoConnection),
-        ):
-            yield self._check_exception, excp, expected
-
-
-class TestZkSharedCacheListEntries(object):
-    """
-    Unit tests for lib.zookeeper.ZkSharedCache._list_entries
-    """
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_sucesss(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get_children"])
-        inst._kazoo.get_children.return_value = [
-            "node0", "node1", "node2", "node3"]
-        # Call
-        ntools.eq_(inst._list_entries(),
-                   {"node0", "node1", "node2", "node3"})
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_no_cache(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get_children"])
-        inst._kazoo.get_children.side_effect = NoNodeError
-        # Call
-        ntools.eq_(inst._list_entries(), set())
-
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def _check_children_exception(self, excp, expected, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._path = "/path"
-        inst._kazoo = create_mock(["get_children"])
-        inst._kazoo.get_children.side_effect = excp
-        # Call
-        ntools.assert_raises(expected, inst._list_entries)
-
-    def test_children_exceptions(self):
-        for excp, expected in (
-            (ConnectionLoss, ZkNoConnection),
-            (SessionExpiredError, ZkNoConnection),
-        ):
-            yield self._check_children_exception, excp, expected
-
-
-class TestZkSharedCacheHandleEntries(object):
-    """
-    Unit test for lib.zookeeper.ZkSharedCache._handle_entries
-    """
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        entry_names = ["entry0", "entry1", "entry2", "entry3"]
-        inst._get = create_mock()
-        inst._get.side_effect = [
-            "data0", ZkNoNodeError, "data2", ZkNoConnection
-        ]
-        inst._path = "/path"
-        inst._handler = create_mock()
-        # Call
-        ntools.eq_(inst._handle_entries(entry_names), 2)
-        # Tests
-        assert_these_calls(inst._get, ([call(i) for i in entry_names]))
-        inst._handler.assert_called_once_with(["data0", "data2"])
-
-
-class TestZkSharedCacheExpire(object):
-    """
-    Unit test for lib.zookeeper.ZkSharedCache.expire
-    """
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_not_connected(self, init):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._zk = create_mock(["is_connected"])
-        inst._zk.is_connected.return_value = False
-        # Call
-        ntools.assert_raises(ZkNoConnection, inst.expire, 42)
-        # Tests
-        inst._zk.is_connected.assert_called_once_with()
-
-    def _setup(self, get_time, entries):
-        inst = ZkSharedCache("zk", "path", "handler")
-        inst._zk = create_mock(["is_connected"])
-        get_time.return_value = 1000
-        inst._entries = entries
-        inst._kazoo = create_mock(["delete"])
-        inst._path = "/path"
-        return inst
-
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def test_success(self, init, get_time):
-        entries = {}
-        for last_seen in 1000, 999, 996, 995, 994, 990, 1001:
-            entries["entry%d" % last_seen] = last_seen
-        inst = self._setup(get_time, entries)
-        # Call
-        inst.expire(5)
-        # Tests
-        assert_these_calls(inst._kazoo.delete, [
-            call("/path/entry994"), call("/path/entry990")
-        ], any_order=True)
-
-    @patch("lib.zookeeper.SCIONTime.get_time", new_callable=create_mock)
-    @patch("lib.zookeeper.ZkSharedCache.__init__", autospec=True,
-           return_value=None)
-    def _check_exception(self, excp, expected, init, get_time):
-        inst = self._setup(get_time, {"entry1": 0})
-        inst._kazoo.delete.side_effect = excp
-        # Call
-        ntools.assert_raises(expected, inst.expire, 5)
-
-    def test_exceptions(self):
-        for excp, expected in (
-            (NoNodeError, ZkNoNodeError),
-            (ConnectionLoss, ZkNoConnection),
-            (SessionExpiredError, ZkNoConnection),
-        ):
-            yield self._check_exception, excp, expected
 
 
 if __name__ == "__main__":
